requirements.txt
attrs==19.3.0
Automat==20.2.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9
cssselect==1.1.0
decorator==4.4.2
future==0.18.2
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
lxml==4.5.0
parsel==1.5.2
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-crontab==2.4.1
python-dateutil==2.8.1
python-telegram-bot==12.6.1
queuelib==1.5.0
requests==2.23.0
schedule==0.6.0
Scrapy==2.0.1
service-identity==18.1.0
six==1.14.0
tornado==6.0.4
Twisted==20.3.0
urllib3==1.25.9
w3lib==1.21.0
zope.interface==5.1.0
attrs==19.3.0
Automat==20.2.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9
cssselect==1.1.0
decorator==4.4.2
future==0.18.2
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
lxml==4.5.0
parsel==1.5.2
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-crontab==2.4.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-telegram-bot==12.6.1
queuelib==1.5.0
requests==2.23.0
schedule==0.6.0
Scrapy==2.0.1
service-identity==18.1.0
six==1.14.0
tornado==6.0.4
Twisted==20.3.0
urllib3==1.25.9
w3lib==1.21.0
zope.interface==5.1.0
